changed radius column type of location table

+++ database/migrations/2020_01_09_145703_create_locations_table.php
@@ -17,7 +17,7 @@ class CreateLocationsTable extends Migration
             $table->bigIncrements('id');
             $table->double("longitude");
             $table->double("latitude");
-            $table->double("radius");
+            $table->float("radius");
             $table->timestamps();
         });
     }
